+++ src/reducers/base/CommonStore.js
@@ -30,6 +30,18 @@ export default class CommonStore {
 
         const items = (state = [], { type, data } = {}) => {
             if (type === INDEX_ITEMS_SUCCEEDED) {
+                /**
+                 * Check if we have a structure like
+                 * {
+                 *  content: [...],
+                 *  last: ...
+                 *  first: ...
+                 * }
+                 */
+                if ('content' in data) {
+                    const {content} = data;
+                    return content;
+                }
                 return data;
             }
             return state;